README.md
The purpose of this project was to model the scoring of runs for the participating teams in the world series by using a Poisson random process. The data was simulated in R programming to plot the trajectory of the Poisson process by specifying parameters and checking the goodness of fit of the data set. The results showed that the higher rate of hits over the rate of homeruns led to higher runs scored.